SparkC-181 backport to 1.6
